backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
# anthropic==0.28.0
google-generativeai
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
aiosqlite==0.20.0
greenlet==3.0.3
aiofiles==23.2.1
tenacity==8.2.3
networkx==3.3

# fastapi==0.111.0
# uvicorn[standard]==0.30.1
# sqlalchemy==2.0.30
# pydantic==2.7.1
# pydantic-settings==2.2.1
# anthropic==0.28.0
# python-dotenv==1.0.1
# httpx==0.27.0
# pytest==8.2.2
# pytest-asyncio==0.23.7
# aiosqlite==0.20.0
# greenlet==3.0.3
# aiofiles==23.2.1
